refactor(team): extract OTLP/usage tracking into broker_otlp_usage.go
Slice 6 of broker.go decomposition. Moves handleUsage,
handleOTLPLogs, the usageEvent type, recordUsageLocked,
applyUsageEvent, usageEventToMessageUsage, cloneMessageUsage,
messageIsWithinUsageAttachWindow, attachUsageToRecentMessagesLocked,
RecordAgentUsage, parseOTLPUsageEvents, nestedMap, otlpAttributesMap,
otlpAnyValue, otlpIntValue, otlpFloatValue, plus the
messageUsageAttachMaxAge constant into a new broker_otlp_usage.go.

Relocates TestParseOTLPUsageEvents, TestBrokerUsageEndpointAggregatesTelemetry,
and TestRecordAgentUsageAttachesToCurrentTurnMessagesOnly from
broker_test.go into broker_otlp_usage_test.go. Adds two new
characterization tests: TestHandleOTLPLogs_RejectsNonPOST and
TestApplyUsageEvent_AccumulatesAcrossCalls.

Total package coverage rises from 63.0% to 63.1%.
